auth: implement login and callback handlers (#8)
Closes https://github.com/sourcegraph/sourcegraph-accounts/issues/43

This PR adds wrapping-handlers that are tailored to work with SAMS:

1. Always use OIDC protocol, as it provide discovery endpoint that only
needs very minimal of the SAMS server configuration, i.e. the URL of the
SAMS instance.
2. The callback handler does one extra thing that generic OIDC handler
should have done: Hit the `/userinfo` endpoint for you, and return a
structured type (`UserInfo`).
3. Allows passthrough of query parameters, `prompt` and `prompt_auth`,
and could be more in the future.

## FAQs

**Why not manage state cookie OOTB like gologin?**

As also mentioned in the added README content, unencrypted cookies are
**strongly discouraged** as they can be tampered with. Although
encrypted cookie is the secure alternative, we do strongly recommend
storing auth state value in a backend component like Redis. Therefore,
function hookers are provided for the service to implement.

<details>
<summary>Resolved</summary>

**Function hookers or interface store?**

I started writing function hookers for managing auth state naturally,
but maybe an interface store would look/work better? 🤔 Open to thoughts.

```go
type StateStore interface {
	SetState(w http.ResponseWriter, r *http.Request, state string) error
	GetState(r *http.Request) (string, error)
	DeleteState(w http.ResponseWriter, r *http.Request)
}
```

</details>



## Test plan

- Tests added for using those handlers to simulate authentication flow
with SAMS.
- Wrote up https://github.com/sourcegraph/cody-analytics/pull/48 to
trial-integrate and tested authentication flow end-to-end.