sams: unify conn config (again), add sams.NewAccountsV1 (#15)
Unified connection config is important for Core Services to operate
clients to e.g. set up private networking. #12 made an attempt at this,
but with #13 I now realize that the first attempt was bad. This change:

1. Makes _only_ `ConnConfig` with SAMS URLs a unified env configuration.
Nothing else really matters, so let's focus on this.
2. Unify root `sams` package to expose constructors for all client
types. Each constructor gets a config struct that has a shared
validation pattern, and embed the shared `sams.ConnConfig`
3. Remove per-client "env constructors" since realistically we don't
care that much about those.

Lots of breaking changes, but we aren't stable-versioned yet, so...
sorry 😅 I'll update `sg/sg` for this

## Test plan

CI