clients/v1: update GetUserRolesResponse to include additional role details (#71)
Updates the GetUserRolesResponse to include additional info about each
role instead of just returning a slice of the fully qualified roles. The
new information includes the `service`, `resource_id` and
`resource_type`.
The `resource_id` and `resource_type` are only included if the role is
granted on a specific resource (i.e. not a service level role)
This will allow clients to query user role info from SAMS to grant
privileged access to specific resources that they are authoritative for.

Closes CORE-399
## Test plan
CI
Ran locally