Disallow SAT in clientcredentials middleware (#101)
SAMS Service Access Tokens function similarly to M2M tokens. Both use
the same introspectToken call to SAMS. For SATs the response includes
the user the token is scoped to.

In the case where a SAT and M2M token both have the same scopes we need
to prevent privilege escalation where a SAT could be supplied to a M2M
RPC which would not be expected to properly enforce the user scoping.

This PR updates the client credentials middleware to fail the request if
an introspected token as a non-empty user field.

Closes CORE-1126
## Test plan
CI

---------

Co-authored-by: ᴊᴏᴇ ᴄʜᴇɴ <jc@unknwon.io>